push all images even for PRs
This will push images with tag :pr-13 where 13 is the PR ID. This will
not update the latest tag. This would make it possible to test images
without having to build them in substratus from Git which can be tricky
when you both modify an image and a base image at the same time.